+++ NpgsqlRestTests/ConnectionTests/ConnectionNameUsingParamsTests.cs
@@ -68,6 +68,6 @@ public class ConnectionNameUsingParamsTests(TestFixture test)
         using var response = await test.Client.GetAsync("/api/get-conn3-connection-name-p/");
         var content = await response.Content.ReadAsStringAsync();
         response?.StatusCode.Should().Be(HttpStatusCode.InternalServerError);
-        content.Should().Be("Connection name conn3 could not be found in options ConnectionStrings dictionary.");
+        content.Should().Be("Connection name conn3 could not be found in options DataSources or ConnectionStrings dictionaries.");
     }
 }